+++ .gitignore
@@ -1,1 +1,2 @@
 data/
+VAE.egg-info/
