+++ VAE/include/custom_parse_rgb.py
